requirements.txt
python==3.9.7
pytorch==1.11.0
numpy==1.21.5
pandas==1.4.1
matplotlib==3.5.2
tqdm==4.63.0
scikit-learn==1.0.2
torch-geometric==2.1.0.post1